Separated ResourceItem and LatestSection, added Swiper, unable to format equal article heights
