+++ react-app/package.json
@@ -15,6 +15,7 @@
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
     "react-redux": "^9.2.0",
+    "swiper": "^11.2.6",
     "tailwind": "^4.0.0"
   },
   "devDependencies": {
